+++ dom_test_utils/package.json
@@ -1,0 +1,8 @@
+{
+	"name": "dom_test_utils",
+	"type": "module",
+	"main": "dist/mod.js",
+	"scripts": {
+		"build": "npx tsc --build"
+	}
+}
