+++ dom_test_utils/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "../tsconfig.json",
+	"compilerOptions": {
+		"declaration": false,
+		"rootDir": "./src",
+		"outDir": "./dist"
+	}
+}
